+++ src/app/(auth)/sign-up/_components/step-complete.tsx
