+++ src/app/(auth)/sign-up/_components/step-password.tsx
