+++ src/app/(auth)/sign-up/_types/funnel.type.ts
